+++ What To Wear/AppDelegate.swift
@@ -58,6 +58,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationWillResignActive(application: UIApplication) {
         // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
         // Use this method to pause ongoing tasks, disable timers, and throttle down OpenGL ES frame rates. Games should use this method to pause the game.
+        if let cameraController = self.window?.rootViewController as? ViewController {
+            cameraController.pauseCaptureSession()
+        }
     }
 
     func applicationDidEnterBackground(application: UIApplication) {
@@ -71,6 +74,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationDidBecomeActive(application: UIApplication) {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+        if let cameraController = self.window?.rootViewController as? ViewController {
+            cameraController.resumeCaptureSession()
+        }
     }
 
     func applicationWillTerminate(application: UIApplication) {